reorganize stuff into usock
