+++ base_squashfs/bin/pyusock
@@ -1,0 +1,2 @@
+#!/bin/sh
+python3 -m usock "$@"
